requirements.txt
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
httpx
apscheduler
python-dotenv
openai
twilio